Part 4: finished 4.20, 4.21, 4.22, 4.23 (Issue #9)

+++ part4/controllers/blogs.js
@@ -1,18 +1,9 @@
-const jwt = require("jsonwebtoken");
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 
-const getTokenFrom = (request) => {
-  const authorization = request.get("authorization");
-  if (authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "");
-  }
-  return null;
-};
-
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
+  const blogs = await Blog.find({}).populate("user", { name: 1, username: 1 });
 
   response.status(200).json(blogs);
 });
@@ -20,17 +11,17 @@ blogsRouter.get("/", async (request, response) => {
 blogsRouter.post("/", async (request, response) => {
   const { title, author, url, likes } = request.body;
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
-  if (!decodedToken.id)
+  if (!request.user) {
     return response.status(401).json({ error: "token invalid" });
+  }
 
-  const user = await User.findById(decodedToken.id);
+  const user = await User.findById(request.user.id);
 
   const blog = new Blog({
-    title,
     author,
-    url,
     likes,
+    title,
+    url,
     user: user.id,
   });
 
@@ -43,7 +34,30 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  if (!request.user) {
+    return response.status(401).json({ error: "invalid token" });
+  }
+
+  const checkBlog = await Blog.findById(request.params.id, { user: 1 });
+
+  if (!checkBlog) {
+    return response.status(404).json({ error: "resource does not exist" });
+  }
+
+  if (request.user.id !== checkBlog.user.toString()) {
+    return response.status(401).json({ error: "not authorised" });
+  }
+
   const deletedBlog = await Blog.findByIdAndRemove(request.params.id);
+
+  if (deletedBlog) {
+    const user = await User.findById(request.user.id);
+    user.blogs = user.blogs.filter(
+      (blog) => blog.toString() !== request.params.id
+    );
+
+    await user.save();
+  }
 
   deletedBlog
     ? response.status(204).end()
@@ -53,11 +67,25 @@ blogsRouter.delete("/:id", async (request, response) => {
 blogsRouter.patch("/:id", async (request, response) => {
   const { title, author, url, likes } = request.body;
 
+  if (!request.user) {
+    return response.status(401).json({ error: "invalid token" });
+  }
+
+  const checkBlog = await Blog.findById(request.params.id, { user: 1 });
+
+  if (!checkBlog) {
+    return response.status(404).json({ error: "resource does not exist" });
+  }
+
+  if (request.user.id !== checkBlog.user.toString()) {
+    return response.status(401).json({ error: "not authorised" });
+  }
+
   const blog = {
-    title,
     author,
-    url,
     likes,
+    title,
+    url,
   };
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
